Refactor tool node executors and helpers
- Introduced _helpers.py to centralize shared helper functions and constants for tool node executors.
- Split LocalExecMixin, MCPMixin, and KwargsResolverMixin into separate files for better organization and maintainability.
- Updated executors to utilize the new helper functions for serialization and status extraction.
- Enhanced error handling and message formatting in the execution flow.
- Changed ExecutionStatus and StopRequestStatus to use StrEnum for better string representation.
- Updated StreamEvent to use StrEnum for consistency.
- Removed deprecated merge_two_state function from utils.py.
