+++ agentflow/core/graph/utils/stream_handler.py
@@ -413,7 +413,7 @@ class StreamHandler[StateT: AgentState](
                 event.data["messages"] = [m.model_dump() for m in messages] if messages else []
                 if messages:
                     lm = messages[-1]
-                    event.content = lm.text() if isinstance(lm.content, list) else lm.content
+                    event.content = str(lm.text() if isinstance(lm.content, list) else lm.content)
                     if isinstance(lm.content, list):
                         event.content_blocks = lm.content
                 event.content_type = [ContentType.STATE, ContentType.MESSAGE]
@@ -563,7 +563,7 @@ class StreamHandler[StateT: AgentState](
             event.data["messages"] = [m.model_dump() for m in messages] if messages else []
             if messages:
                 fm = messages[-1]
-                event.content = fm.text() if isinstance(fm.content, list) else fm.content
+                event.content = str(fm.text() if isinstance(fm.content, list) else fm.content)
                 if isinstance(fm.content, list):
                     event.content_blocks = fm.content
             event.content_type = [ContentType.STATE, ContentType.MESSAGE]
